+++ Source/Whistle/IWhistle.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Narkhedegs.Diagnostics
+namespace Narkhedegs
 {
     /// <summary>
     /// Provides methods to start external executable from within a .NET process.