+++ Tests/Whistle.Tests/Helpers/WhistleOptionsGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Narkhedegs.Diagnostics.Tests.Helpers
+namespace Narkhedegs.Tests.Helpers
 {
     public static class WhistleOptionsGenerator
     {