+++ Tests/Whistle.Tests/when_constructing_a_whistle.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using Moq;
-using Narkhedegs.Diagnostics.Tests.Helpers;
+using Narkhedegs.Tests.Helpers;
 using NUnit.Framework;
 
-namespace Narkhedegs.Diagnostics.Tests
+namespace Narkhedegs.Tests
 {
     [TestFixture]
     public class when_constructing_a_whistle