+++ ReleaseNotes.md
@@ -3,3 +3,4 @@
 
 ### New in 0.0.2 (Released 2015/09/15)
 * Assign correct values to CreateNoWindow, ErrorDialog and WindowStyle properties of ProcessStartInfo
+* Implement functionality to add Environment Variables to the external executable