+++ ReleaseNotes.md
@@ -4,3 +4,6 @@
 ### New in 0.0.2 (Released 2015/09/15)
 * Assign correct values to CreateNoWindow, ErrorDialog and WindowStyle properties of ProcessStartInfo
 * Implement functionality to add Environment Variables to the external executable
+
+### New in 0.0.3 (Released 2015/09/16)
+* Change namespace from Narkhedegs.Diagnostics to Narkhedegs